requirements.txt
Django==1.6.4
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.5
gunicorn==0.16.1
psycopg2==2.5.2
pystache==0.5.3
static==1.0.2
wsgiref==0.1.2
django-dajax==0.9.2
django-dajaxice==0.6